# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a projectc-custodial-wallet.",
        "title": "Swagger projectc-custodial-wallet",
        "contact": {
            "name": "guyuxiang",
            "url": "https://guyuxiang.github.io",
            "email": "gu_yuxiang@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/admin/signature/key/upsert": {
            "post": {
                "description": "Persist or update a signature key pair by publickeyId.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SignatureKey"
                ],
                "summary": "UpsertSignatureKey",
                "parameters": [
                    {
                        "description": "Signature key upsert request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SignatureKeyUpsertRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "models.SignatureKeyUpsertRequest": {
            "type": "object",
            "required": [
                "privateKey",
                "publicKey",
                "publickeyId"
            ],
            "properties": {
                "privateKey": {
                    "type": "string"
                },
                "publicKey": {
                    "type": "string"
                },
                "publickeyId": {
                    "type": "string"
                }
            }
        },
...
    }
}